Extract out benchmarks into its own crate
