+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
 	'wnfs',
+	'wnfs-bench',
 	'wnfs-wasm'
 ]
