+++ wnfs-bench/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "wnfs-bench"
+version = "0.1.9"
+description = "WNFS Benchmarks"
+keywords = ["wnfs", "webnative", "ipfs", "decentralisation", "benchmarks"]
+categories = ["filesystem", "cryptography", "web-programming", "wasm"]
+edition = "2021"
+repository = "https://github.com/WebNativeFileSystem/rs-wnfs/tree/main/bench"
+homepage = "https://fission.codes"
+authors = ["The Fission Authors"]
+
+[dependencies]
+proptest = "1.0"
+wnfs = { path = "../wnfs", optional = true }
+async-std = { version = "1.11", features = ["attributes"] }
+
+[dev-dependencies]
+criterion = { version = "0.4", features = ["async_std"] }
+
+[lib]
+path = "lib.rs" # TODO(appcypher)
+
+[features]
+default = ["wnfs/test_strategies"]
+
+[[bench]]
+name = "hamt"
+harness = false
+path = "hamt.rs"
+
+[[bench]]
+name = "namefilter"
+harness = false
+path = "namefilter.rs"
