+++ wnfs/Cargo.toml
@@ -40,11 +40,11 @@ sha3 = "0.10"
 skip_ratchet = { version = "0.1", features = ["serde"] }
 thiserror = "1.0"
 xxhash-rust = { version = "0.8", features = ["xxh3"] }
+proptest = { version = "1.0", optional = true }
 
 [dev-dependencies]
 criterion = { version = "0.4", features = ["async_std"] }
 env_logger = "0.9"
-proptest = "1.0"
 rand = "0.8"
 test-log = "0.2"
 test-strategy = "0.2"
@@ -57,11 +57,4 @@ bench = false
 [features]
 default = []
 wasm = []
-
-[[bench]]
-name = "hamt"
-harness = false
-
-[[bench]]
-name = "namefilter"
-harness = false
+test_strategies = ["proptest"]
