+++ wnfs/private/hamt/strategies.rs
@@ -1,0 +1,165 @@
+use crate::{private::hamt::node::*, BlockStore};
+use anyhow::Result;
+use proptest::{collection::*, prelude::*, strategy::Shuffleable};
+use serde::{de::DeserializeOwned, Serialize};
+use std::{collections::HashMap, fmt::Debug, hash::Hash, rc::Rc};
+
+#[derive(Debug, Clone)]
+pub enum Operation<K, V> {
+    Insert(K, V),
+    Remove(K),
+}
+
+impl<K, V> Operation<K, V> {
+    pub fn can_be_swapped_with(&self, other: &Operation<K, V>) -> bool
+    where
+        K: PartialEq,
+        V: PartialEq,
+    {
+        match (self, other) {
+            (Operation::Insert(key_a, val_a), Operation::Insert(key_b, val_b)) => {
+                // We can't swap if the keys are the same and values different.
+                // Because in those cases operation order matters.
+                // E.g. insert "a" 10, insert "a" 11 != insert "a" 11, insert "a" 10
+                // But insert "a" 10, insert "b" 11 == insert "b" 11, insert "a" 10
+                // Or insert "a" 10, insert "a" 10 == insert "a" 10, insert "a" 10 ('swapped')
+                key_a != key_b || val_a == val_b
+            }
+            (Operation::Insert(key_i, _), Operation::Remove(key_r)) => {
+                // We can only swap if these two operations are unrelated.
+                // Otherwise order matters.
+                // E.g. insert "a" 10, remove "a" != remove "a", insert "a" 10
+                key_i != key_r
+            }
+            (Operation::Remove(key_r), Operation::Insert(key_i, _)) => {
+                // same as above
+                key_i != key_r
+            }
+            (Operation::Remove(_), Operation::Remove(_)) => {
+                // Removes can always be swapped
+                true
+            }
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct Operations<K, V>(Vec<Operation<K, V>>);
+
+impl<K: PartialEq, V: PartialEq> Shuffleable for Operations<K, V> {
+    fn shuffle_len(&self) -> usize {
+        self.0.len()
+    }
+
+    /// Swaps the values if that wouldn't change the semantics.
+    /// Otherwise it's a no-op.
+    fn shuffle_swap(&mut self, a: usize, b: usize) {
+        use std::cmp;
+        if a == b {
+            return;
+        }
+        let min = cmp::min(a, b);
+        let max = cmp::max(a, b);
+        let left = &self.0[min];
+        let right = &self.0[max];
+
+        for i in min..=max {
+            let neighbor = &self.0[i];
+            if !left.can_be_swapped_with(neighbor) {
+                return;
+            }
+            if !right.can_be_swapped_with(neighbor) {
+                return;
+            }
+        }
+
+        // The reasoning for why this works now, is following:
+        // Let's look at an example. We checked that we can do all of these swaps:
+        // a x y z b
+        // x a y z b
+        // x y a z b
+        // x y z a b
+        // x y z b a
+        // x y b z a
+        // x b y z a
+        // b x y z a
+        // Observe how a moves to the right
+        // and b moves to the left.
+        // The end result is the same as
+        // just swapping a and b.
+        // With all calls to `can_be_swapped_with` above
+        // we've made sure that this operation is now safe.
+
+        self.0.swap(a, b);
+    }
+}
+
+pub async fn node_from_operations<K, V, B: BlockStore>(
+    operations: Operations<K, V>,
+    store: &mut B,
+) -> Result<Rc<Node<K, V>>>
+where
+    K: DeserializeOwned + Serialize + Clone + Debug + AsRef<[u8]>,
+    V: DeserializeOwned + Serialize + Clone + Debug,
+{
+    let mut node: Rc<Node<K, V>> = Rc::new(Node::default());
+    for op in operations.0 {
+        match op {
+            Operation::Insert(key, value) => {
+                node = node.set(key.clone(), value, store).await?;
+            }
+            Operation::Remove(key) => {
+                (node, _) = node.remove(&key, store).await?;
+            }
+        };
+    }
+
+    Ok(node)
+}
+
+pub fn hash_map_from_operations<K: Debug + Clone + Hash + Eq, V: Debug + Clone + Eq>(
+    operations: Operations<K, V>,
+) -> HashMap<K, V> {
+    let mut map = HashMap::default();
+    for op in operations.0 {
+        match op {
+            Operation::Insert(key, value) => {
+                map.insert(key, value);
+            }
+            Operation::Remove(key) => {
+                map.remove(&key);
+            }
+        }
+    }
+    map
+}
+
+pub fn operation<K: Debug, V: Debug>(
+    key: impl Strategy<Value = K>,
+    value: impl Strategy<Value = V>,
+) -> impl Strategy<Value = Operation<K, V>> {
+    (any::<bool>(), key, value).prop_map(|(is_insert, key, value)| {
+        if is_insert {
+            Operation::Insert(key, value)
+        } else {
+            Operation::Remove(key)
+        }
+    })
+}
+
+pub fn operations<K: Debug, V: Debug>(
+    key: impl Strategy<Value = K>,
+    value: impl Strategy<Value = V>,
+    size: impl Into<SizeRange>,
+) -> impl Strategy<Value = Operations<K, V>> {
+    vec(operation(key, value), size).prop_map(|vec| Operations(vec))
+}
+
+pub fn operations_and_shuffled<K: PartialEq + Clone + Debug, V: PartialEq + Clone + Debug>(
+    key: impl Strategy<Value = K>,
+    value: impl Strategy<Value = V>,
+    size: impl Into<SizeRange>,
+) -> impl Strategy<Value = (Operations<K, V>, Operations<K, V>)> {
+    operations(key, value, size)
+        .prop_flat_map(|operations| (Just(operations.clone()), Just(operations).prop_shuffle()))
+}
